# old_python_fastapi_backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SafeRoute API",
    "description": "SafeRoute finds safer routes through Southampton by analyzing UK Police crime data.\n\nFeatures: Safety heatmaps with crime risk visualization, multiple route alternatives ranked by safety score, historical crime analytics with time-of-day weighting, Redis caching, and automated monthly data updates.\n\nData sources: UK Police API for crime data, OpenRouteService for routing, H3 hexagonal spatial indexing for efficient geographic queries.\n\nAuthentication is optional. Authenticated users get route history tracking, custom preferences, and higher rate limits. Use /auth/login to obtain a JWT access token.\n    ",
    "contact": {
      "name": "SafeRoute API Support",
      "url": "https://github.com/ss2d22/saferoute"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Register a new user.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Login user and return tokens.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "Refresh access token using refresh token.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Logout user by revoking refresh token.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Current User",
        "description": "Get current authenticated user info.",
        "operationId": "get_current_user_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/routes/safe": {
      "post": {
        "tags": [
          "routes"
        ],
        "summary": "Get safety-scored route options",
        "description": "Get route alternatives with safety scores.\n\n    Queries OpenRouteService for up to 3 route options between two points, then\n    analyzes each route against historical crime data to calculate safety scores.\n\n    **Features:**\n    - Multiple route alternatives with safety scoring\n    - ~100m segment analysis for detailed risk assessment\n    - Crime hotspot detection along routes\n    - Recency-weighted crime data\n    - Time-of-day crime pattern adjustment\n    - Route history tracking (authenticated users)\n    - Cached for performance\n\n    **Transport Modes:**\n    - `foot-walking`: Walking routes\n    - `cycling-regular`: Bicycle routes\n    - `driving-car`: Car routes",
        "operationId": "get_safe_routes_api_v1_routes_safe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SafeRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully calculated route alternatives with safety scores",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeRouteResponse"
                }
              }
            }
          },
          "404": {
            "description": "No routes found between the specified points",
            "content": {
              "application/json": {
                "example": {
                  "detail": "No routes found"
                }
              }
            }
          },
          "500": {
            "description": "Error calculating routes or fetching crime data"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/routes/{route_id}": {
      "get": {
        "tags": [
          "routes"
        ],
        "summary": "Get Route",
        "description": "Get cached route by ID.\n\nNote: Route caching not yet implemented.",
        "operationId": "get_route_api_v1_routes__route_id__get",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Coordinate": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "Coordinate",
        "description": "Geographic coordinate."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "User login request."
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "description": "Refresh token request."
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "description": "User registration request."
      },
...
      "RouteHotspot": {
        "properties": {
          "segment_index": {
            "type": "integer",
            "title": "Segment Index"
          },
          "location": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Location"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "segment_index",
          "location",
          "risk_score",
          "risk_level",
          "description"
        ],
        "title": "RouteHotspot",
        "description": "Crime hotspot along route."
      },
      "RoutePreferences": {
        "properties": {
          "safety_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Safety Weight",
            "default": 0.8
          },
          "lookback_months": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 3.0,
            "title": "Lookback Months",
            "default": 12
          },
          "time_of_day_sensitive": {
            "type": "boolean",
            "title": "Time Of Day Sensitive",
            "default": true
          },
          "category_weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Weights"
          }
        },
        "type": "object",
        "title": "RoutePreferences",
        "description": "Route safety scoring preferences."
      },
      "RouteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "is_recommended": {
            "type": "boolean",
            "title": "Is Recommended"
          },
          "safety_score": {
            "type": "number",
            "title": "Safety Score"
          },
          "risk_class": {
            "type": "string",
            "title": "Risk Class"
          },
          "distance_m": {
            "type": "integer",
            "title": "Distance M"
          },
          "duration_s": {
            "type": "integer",
            "title": "Duration S"
          },
          "geometry": {
            "additionalProperties": true,
            "type": "object",
            "title": "Geometry"
          },
          "instructions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Instructions",
            "default": []
          },
          "stats": {
            "$ref": "#/components/schemas/RouteStats"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rank",
          "is_recommended",
          "safety_score",
          "risk_class",
          "distance_m",
          "duration_s",
          "geometry",
          "stats"
        ],
        "title": "RouteResponse",
        "description": "Single route response."
      },
      "RouteSegment": {
        "properties": {
          "segment_index": {
            "type": "integer",
            "title": "Segment Index"
          },
          "start_point": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Start Point"
          },
          "end_point": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "End Point"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "cell_count": {
            "type": "integer",
            "title": "Cell Count"
          }
        },
        "type": "object",
        "required": [
          "segment_index",
          "start_point",
          "end_point",
          "risk_score",
          "cell_count"
        ],
        "title": "RouteSegment",
        "description": "Individual route segment safety data."
      },
      "RouteStats": {
        "properties": {
          "total_weighted_risk": {
            "type": "number",
            "title": "Total Weighted Risk"
          },
          "max_segment_risk": {
            "type": "number",
            "title": "Max Segment Risk"
          },
          "avg_segment_risk": {
            "type": "number",
            "title": "Avg Segment Risk",
            "default": 0.0
          },
          "segment_count": {
            "type": "integer",
            "title": "Segment Count",
            "default": 0
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/RouteSegment"
            },
            "type": "array",
            "title": "Segments",
            "default": []
          },
          "hotspots": {
            "items": {
              "$ref": "#/components/schemas/RouteHotspot"
            },
            "type": "array",
            "title": "Hotspots",
            "default": []
          },
          "crime_breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Crime Breakdown",
            "default": {}
          },
          "cells_analyzed": {
            "type": "integer",
            "title": "Cells Analyzed",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "total_weighted_risk",
          "max_segment_risk"
        ],
        "title": "RouteStats",
        "description": "Route safety statistics."
      },
      "SafeRouteRequest": {
        "properties": {
          "origin": {
            "$ref": "#/components/schemas/Coordinate"
          },
          "destination": {
            "$ref": "#/components/schemas/Coordinate"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "foot-walking"
          },
          "departure_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Time"
          },
          "preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutePreferences"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "origin",
          "destination"
        ],
        "title": "SafeRouteRequest",
        "description": "Request for safe routes."
      },
      "SafeRouteResponse": {
        "properties": {
          "routes": {
            "items": {
              "$ref": "#/components/schemas/RouteResponse"
            },
            "type": "array",
            "title": "Routes"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "routes",
          "meta"
        ],
        "title": "SafeRouteResponse",
        "description": "Response with multiple safe route options."
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "description": "Token response."
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "is_active",
          "settings",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "User info response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}